avm1: Move all actions from Avm1 to a new StackFrame
